부호 바꾸는 버튼 클릭이벤트 생성, 정수부분에만 천단위 구분 기호 표시하는 코드 메서드로 만듬

+++ WinCalcPro/WinCalcPro/StandardCalc.cs
@@ -166,7 +166,7 @@ namespace WinCalcPro
         /// </summary>
         /// <param name="x">입력값</param>
         /// <returns>부호를 바꿔서 반환</returns>
-        public double ChangeSign(double x)
+        public decimal ChangeSign(decimal x)
         {
             return -x; // 부호를 바꿔서 반환
         }
